requirements.txt
einops==0.6.1
h5py==3.9.0
numpy==1.23.5
openai==0.27.8
Pillow==9.3.0
torch==1.13.1
transformers==4.22.0
